requirements.txt
absl-py==1.0.0
altgraph==0.17.2
asgiref==3.5.2
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
distlib==0.3.4
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt== 5.2.0
docutils==0.18.1
filelock==3.6.0
fonttools==4.32.0
html5lib==1.1
idna==3.3
imutils==0.5.4
kiwisolver==1.4.2
matplotlib==3.5.1
packaging==21.3
Pillow==9.1.0
platformdirs==2.5.1
protobuf==3.20.0
psycopg2-binary==2.9.3
Pygments==2.11.2
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
urllib3==1.26.9
virtualenv==20.14.1
webencodings==0.5.1
